[Backend] REST endpoints implement #7 - PostController endpointok elkészültek - Adatbázis sémában egy apróbb javítás be lett téve, hogy működjön az adatbázis <-> JPA id generálás - Exception kezelés alapjai elkészültek.

+++ src/main/java/hu/dominikvaradi/sociallybackend/flows/comment/domain/dto/CommentResponseDto.java
@@ -19,5 +19,5 @@ public class CommentResponseDto {
 	private UUID authorId;
 	private String authorName;
 	private LocalDateTime created;
-	EnumMap<Reaction, Integer> reactionsCount;
+	EnumMap<Reaction, Long> reactionsCount;
 }
